🚸 No longer raise error when notebook not saved

+++ lamindb/_finish.py
@@ -3,10 +3,12 @@ from __future__ import annotations
 import builtins
 import re
 from datetime import datetime, timezone
+from time import sleep
 from typing import TYPE_CHECKING
 
 import lamindb_setup as ln_setup
 from lamin_utils import logger
+from lamin_utils._logger import LEVEL_TO_COLORS, LEVEL_TO_ICONS, RESET_COLOR
 from lamindb_setup.core.hashing import hash_file
 
 from lamindb.core.exceptions import NotebookNotSaved
@@ -19,7 +21,59 @@ if TYPE_CHECKING:
 
 
 def get_save_notebook_message() -> str:
-    return f"Please save the notebook in your editor (shortcut `{get_shortcut()}`) within 2 sec before calling `finish()`"
+    return f"please save the notebook in your editor (shortcut `{get_shortcut()}`)"
+
+
+# this code was originally in nbproject by the same authors
+def check_consecutiveness(
+    nb, calling_statement: str = None, silent_success: bool = True
+) -> bool:
+    """Check whether code cells have been executed consecutively.
+
+    Needs to be called in the last code cell of a notebook.
+    Otherwise raises `RuntimeError`.
+
+    Returns cell transitions that violate execution at increments of 1 as a list
+    of tuples.
+
+    Args:
+        nb: Notebook content.
+        calling_statement: The statement that calls this function.
+    """
+    cells = nb.cells
+
+    violations = []
+    prev = 0
+
+    ccount = 0  # need to initialize because notebook might note have code cells
+    # and below, we check if ccount is None
+    for cell in cells:
+        cell_source = "".join(cell["source"])
+        if cell["cell_type"] != "code" or cell_source == "":
+            continue
+
+        if calling_statement is not None and calling_statement in cell_source:
+            continue
+
+        ccount = cell["execution_count"]
+        if ccount is None or prev is None or ccount - prev != 1:
+            violations.append((prev, ccount))
+
+        prev = ccount
+
+    # ignore the very last code cell of the notebook
+    # `check_consecutiveness` is being run during publish if `last_cell`` is True
+    # hence, that cell has ccount is None
+    if ccount is None:
+        violations.pop()
+
+    any_violations = len(violations) > 0
+    if any_violations:
+        logger.warning(f"cells {violations} were not run consecutively")
+    elif not silent_success:
+        logger.success("cell execution numbers increase consecutively")
+
+    return not any_violations
 
 
 def get_shortcut() -> str:
@@ -170,12 +224,40 @@ def save_context_core(
     is_r_notebook = filepath.suffix in {".qmd", ".Rmd"}
     source_code_path = filepath
     report_path: Path | None = None
-    # for notebooks, we need more work
-    if is_ipynb:
+    save_source_code_and_report = True
+    if is_run_from_ipython:  # python notebooks in interactive session
+        import nbproject
+
+        # it might be that the user modifies the title just before ln.finish()
+        if (nbproject_title := nbproject.meta.live.title) != transform.description:
+            transform.description = nbproject_title
+            transform.save()
+        if not ln_setup._TESTING:
+            save_source_code_and_report = False  # init to False
+            for repeat in range(8):
+                if get_seconds_since_modified(filepath) > 2.5:
+                    if repeat == 0:
+                        prefix = (
+                            f"{LEVEL_TO_COLORS[20]}{LEVEL_TO_ICONS[20]}{RESET_COLOR}"
+                        )
+                        print(f"{prefix} {get_save_notebook_message()}", end=" ")
+                    elif repeat == 7:
+                        print(".", end="\n")
+                    else:
+                        print(".", end="")
+                    sleep(1)
+                else:
+                    if repeat > 0:
+                        prefix = (
+                            f"{LEVEL_TO_COLORS[25]}{LEVEL_TO_ICONS[25]}{RESET_COLOR}"
+                        )
+                        print(f" {prefix}")
+                    save_source_code_and_report = True
+                    break
+    if is_ipynb:  # could be from CLI outside interactive session
         try:
             import jupytext  # noqa: F401
             from nbproject.dev import (
-                check_consecutiveness,
                 read_notebook,
             )
         except ImportError:
@@ -188,7 +270,9 @@ def save_context_core(
             )
             if not is_consecutive:
                 response = "n"  # ignore_non_consecutive == False
-                if ignore_non_consecutive is None:
+                if ignore_non_consecutive is None:  # only print warning
+                    response = "y"  # we already printed the warning
+                else:  # ask user to confirm
                     response = input(
                         "   Do you still want to proceed with finishing? (y/n) "
                     )
@@ -213,43 +297,35 @@ def save_context_core(
             logger.warning(
                 f"no html report found; to attach one, create an .html export for your {filepath.suffix} file and then run: lamin save {filepath}"
             )
-    if is_run_from_ipython:  # python notebooks
-        import nbproject
-
-        # it might be that the user modifies the title just before ln.finish()
-        if (nbproject_title := nbproject.meta.live.title) != transform.description:
-            transform.description = nbproject_title
-            transform.save()
-        if get_seconds_since_modified(filepath) > 2 and not ln_setup._TESTING:
-            raise NotebookNotSaved(get_save_notebook_message())
     if report_path is not None and not from_cli:  # R notebooks
         if get_seconds_since_modified(report_path) > 2 and not ln_setup._TESTING:
             # this can happen when auto-knitting an html with RStudio
             raise NotebookNotSaved(get_save_notebook_message())
     ln.settings.creation.artifact_silence_missing_run_warning = True
-    # track source code
-    hash, _ = hash_file(source_code_path)  # ignore hash_type for now
-    if transform.hash is not None:
-        # check if the hash of the transform source code matches
-        # (for scripts, we already run the same logic in track() - we can deduplicate the call at some point)
-        if hash != transform.hash:
-            response = input(
-                f"You are about to overwrite existing source code (hash '{transform.hash}') for Transform('{transform.uid}')."
-                f" Proceed? (y/n)"
-            )
-            if response == "y":
-                transform.source_code = source_code_path.read_text()
-                transform.hash = hash
+    # save source code
+    if save_source_code_and_report:
+        hash, _ = hash_file(source_code_path)  # ignore hash_type for now
+        if transform.hash is not None:
+            # check if the hash of the transform source code matches
+            # (for scripts, we already run the same logic in track() - we can deduplicate the call at some point)
+            if hash != transform.hash:
+                response = input(
+                    f"You are about to overwrite existing source code (hash '{transform.hash}') for Transform('{transform.uid}')."
+                    f" Proceed? (y/n)"
+                )
+                if response == "y":
+                    transform.source_code = source_code_path.read_text()
+                    transform.hash = hash
+                else:
+                    logger.warning("Please re-run `ln.track()` to make a new version")
+                    return "rerun-the-notebook"
             else:
-                logger.warning("Please re-run `ln.track()` to make a new version")
-                return "rerun-the-notebook"
+                logger.debug("source code is already saved")
         else:
-            logger.debug("source code is already saved")
-    else:
-        transform.source_code = source_code_path.read_text()
-        transform.hash = hash
+            transform.source_code = source_code_path.read_text()
+            transform.hash = hash
 
-    # track environment
+    # track run environment
     if run is not None:
         env_path = ln_setup.settings.cache_dir / f"run_env_pip_{run.uid}.txt"
         if env_path.exists():
@@ -282,43 +358,45 @@ def save_context_core(
         save_run_logs(run)
 
     # track report and set is_consecutive
-    if run is not None:
-        if report_path is not None:
-            if is_r_notebook:
-                title_text, report_path = clean_r_notebook_html(report_path)
-                if title_text is not None:
-                    transform.description = title_text
-            if run.report_id is not None:
-                hash, _ = hash_file(report_path)  # ignore hash_type for now
-                if hash != run.report.hash:
-                    response = input(
-                        f"You are about to overwrite an existing report (hash '{run.report.hash}') for Run('{run.uid}'). Proceed? (y/n)"
-                    )
-                    if response == "y":
-                        run.report.replace(report_path)
-                        run.report.save(upload=True, print_progress=False)
+    if save_source_code_and_report:
+        if run is not None:
+            if report_path is not None:
+                if is_r_notebook:
+                    title_text, report_path = clean_r_notebook_html(report_path)
+                    if title_text is not None:
+                        transform.description = title_text
+                if run.report_id is not None:
+                    hash, _ = hash_file(report_path)  # ignore hash_type for now
+                    if hash != run.report.hash:
+                        response = input(
+                            f"You are about to overwrite an existing report (hash '{run.report.hash}') for Run('{run.uid}'). Proceed? (y/n)"
+                        )
+                        if response == "y":
+                            run.report.replace(report_path)
+                            run.report.save(upload=True, print_progress=False)
+                        else:
+                            logger.important("keeping old report")
                     else:
-                        logger.important("keeping old report")
+                        logger.important("report is already saved")
                 else:
-                    logger.important("report is already saved")
-            else:
-                report_file = ln.Artifact(
-                    report_path,
-                    description=f"Report of run {run.uid}",
-                    _branch_code=0,  # hidden file
-                    run=False,
+                    report_file = ln.Artifact(
+                        report_path,
+                        description=f"Report of run {run.uid}",
+                        _branch_code=0,  # hidden file
+                        run=False,
+                    )
+                    report_file.save(upload=True, print_progress=False)
+                    run.report = report_file
+                if is_r_notebook:
+                    # this is the "cleaned" report
+                    report_path.unlink()
+                logger.debug(
+                    f"saved transform.latest_run.report: {transform.latest_run.report}"
                 )
-                report_file.save(upload=True, print_progress=False)
-                run.report = report_file
-            if is_r_notebook:
-                # this is the "cleaned" report
-                report_path.unlink()
-            logger.debug(
-                f"saved transform.latest_run.report: {transform.latest_run.report}"
-            )
-        run._is_consecutive = is_consecutive
+            run._is_consecutive = is_consecutive
 
-        # save both run & transform records if we arrive here
+    # save both run & transform records if we arrive here
+    if run is not None:
         run.save()
     transform.save()
 
@@ -330,19 +408,31 @@ def save_context_core(
         hours = seconds // 3600
         minutes = (seconds % 3600) // 60
         secs = seconds % 60
-        formatted_run_time = f"{days}d {hours}h {minutes}m {secs}s"
+        formatted_run_time = (
+            f"{days}d"
+            if days != 0
+            else "" + f"{hours}h"
+            if hours != 0
+            else "" + f"{minutes}m"
+            if minutes != 0
+            else "" + f"{secs}s"
+        )
 
         logger.important(
             f"finished Run('{run.uid[:8]}') after {formatted_run_time} at {format_field_value(run.finished_at)}"
         )
     if ln_setup.settings.instance.is_on_hub:
-        identifier = ln_setup.settings.instance.slug
+        instance_slug = ln_setup.settings.instance.slug
         logger.important(
-            f"go to: https://lamin.ai/{identifier}/transform/{transform.uid}"
+            f"go to: https://lamin.ai/{instance_slug}/transform/{transform.uid}"
         )
-        if not from_cli:
+        if not from_cli and save_source_code_and_report:
             thing = "notebook" if (is_ipynb or is_r_notebook) else "script"
             logger.important(
-                f"if you want to update your {thing} without re-running it, use `lamin save {filepath}`"
+                f"if you want to update your {thing} from the CLI, run: lamin save {filepath}"
             )
+    if not save_source_code_and_report:
+        logger.warning(
+            f"did not save source code and report, to do so, run: lamin save {filepath}"
+        )
     return None
